Quiet datacompy (#2087)
Fixes #2084
The 5 warnings shown on `import ribasim` were either logging warnings or
UserWarnings, so both need to be suppressed.

I had to add `noqa: E402` because we now have logging code above our
imports. But if I move that line down it is too late and the warnings
are logged.