Modify QGIS style symlinking (#2093)
This changes the style symlinking a bit from how it it was implemented
in #1713.
I ran into permission denied errors locally on
`styles_source_path.unlink` and this fixes them for me.

There are 3 places for styles:
1. ribasim python has them in the repo and uses them for adding styles
on model.write
2. ribasim_qgis expects them as a default style fallback and has them in
gitignore
3. qgis_env is such that `pixi run qgis` directly starts with the
important plugins loaded

With this PR, 3 points to 2, and 2 points to 1.

After this is merged we need to double check if ribasim_qgis.zip still
contains the styles under core/styles.