Add cyclic forcing support (#2081)
Fixes https://github.com/Deltares/Ribasim/issues/153.

Notes:
- With timeseries which are extrapolated in a periodic manner, the first
and last data values must be validated to be the same. For linear
interpolations this ensures continuity, and for constant interpolation
this value would be otherwise ignored.
- Time periods like years and months do not always have the same length,
so this can get out of sync when these time periods are repeated. This
should at least be documented well, and might need a follow-up issue.
- We add all time values in the input to the `tstops`. With cycling
forcing this has to be expanded to all time values in all periods that
fall within the simulation period.