CI: Force rebuilding on tags (#2152)
Teamcity re-uses builds in a build chain when possible, so once the tag
triggered the release chain, it re-used the previously build
executable(s) for the commit of the tag. We can disable this by
`reuseBuilds = ReuseBuilds.NO` in the specific snapshot builds.

This also enables us to skip builds on Teamcity (and already on Github)
with a commit message that starts with `[skip ci]`. I've documented this
as well.

Finally, I noted we create duplicate builds, once for a(ny) branch, and
once for a created PR. This is bad (especially because I disabled some
re-use above), so I've added some branchfilter to the default branch
only (as [advised when using PR
integration](https://www.jetbrains.com/help/teamcity/pull-requests.html#Interaction+with+VCS+Roots)).
I also added this PR integration to the testmodel build which didn't
have it yet.

Fixes #2040 
Fixes #2128