Add `model.to_fews(region_dir)` (#2161)
This adds a function `model.to_fews(region_dir)` that converts the
network and results to files that Delft-FEWS can directly handle. It is
marked as experimental for now.

@gijsber is working on a Delft-FEWS configuration that can be used to
visualize model results, to complement our existing tools. We'll likely
add this configuration to this monorepo since it is generic. #2159 also
pertains to this work.

What is especially nice is the spatio-temporal support of Delft-FEWS, so
we can make visualizations like this:


![image](https://github.com/user-attachments/assets/2e61bf82-0d7d-4558-a645-755d7e763b74)

In theory we can support similar functionality with QGIS, but looking at
the plots in https://github.com/Deltares/Ribasim/pull/1369 this would
likely need work in QGIS itself. So this is really a quick win to be
able to inspect models better.

---------

Co-authored-by: Maarten Pronk <git@evetion.nl>