Use `autodiff = True` as the default again (#2183)
After https://github.com/Deltares/Ribasim/issues/2134 I wanted to check
if https://github.com/Deltares/Ribasim/issues/1834 was fixed, and it
was, so this essentially reverts
https://github.com/Deltares/Ribasim/issues/1832.

Making it the default since it is noticably faster and should be more
accurate as well. I've seen it do better with convergence bottlenecks as
well.

```julia
# hws_2025_3_0
# autodiff=false
# 57.718218 seconds (15.07 M allocations: 10.536 GiB, 2.00% gc time, 2 lock conflicts)
# autodiff=true
# 51.234467 seconds (14.27 M allocations: 10.210 GiB, 1.76% gc time, 2 lock conflicts)

# AmstelGooienVecht_parameterized_2025_3_4
# autodiff=false
# 7.283102 seconds (2.52 M allocations: 1.033 GiB, 2.59% gc time, 6 lock conflicts)
# autodiff=true
# 5.603329 seconds (951.18 k allocations: 847.611 MiB, 2.76% gc time, 5 lock conflicts)
```