Simplify "Pre-commit auto update" (#2185)
Using the current docs:
https://github.com/browniebroke/github-actions#pre-commit-auto-update
All these other steps are already part of 

https://github.com/browniebroke/github-actions/blob/bcdb59e937d32c481e5ea2233aff21f8053a934a/.github/workflows/pre-commit-autoupdate.yml

I also added a commit that sets the hook version back and ruins some
formatting so we can test if it fixes this automatically. Because the
main problem with this Action was that a new e.g. ruff version would
change the formatting, and this got rejected by the new pre-commit. That
caused the auto update job to fail and never appear.