Allow max 1 outflow neighbour for FlowBoundary
Fixes #2072. It was possible to let one FlowBoundary flow into multiple Basins. This is no longer allowed, so this is a breaking change. As far as I know it wasn't used though. It is quite confusing to have multiple outflow neighbours, because a Q at the FlowBoundary that goes to n Basins then becomes a nQ total inflow. And usually a FlowBoundary represents a single physical inflow point, that should always go to one point only.

The LevelBoundary can still be connected via resistance nodes to multiple Basins, but this makes sense when representing e.g. the sea.