Add temporal Flow and Basin results to QGIS. (#2208)
And adds "show topology" button on the layer context menu.

Fixes #2178
Fixes #2171 

This creates a new Results group in QGIS for each loaded model, and adds
a BasinNode and FlowLink layer to it. These are duplicates of the Node
and Link layer for the result data (and thus only keeping
`node_type=Basin`, and `link_type=flow`).

Result arrow files (basin, flow and concentration) are read with pandas
and indexed on time (concentration is widened first) and saved on the
widget. Fields on the BasinNode and FlowLink are created to store the
result variables. A specific time slice, by default the last, or by a
signal from the temporal manager is then used to update these fields.
The temporal manager is initialized based on the start/endtime and
timestep from .toml file.

In effect you now have Node and Link tables with all result variables
for a given timestep, meaning you're free to add symbology/labels as you
would for any (static) layer.

The layer context menu is expanded to show a "Show topology" checkmark,
which sets a layer variable, which in turn is used by the layer
symbology to either show `@geometry` or
`make_line(start_point(@geometry), end_point( @geometry))`.

---

For @huite and @JoerivanEngelen, updating the (memory) layers on the
fly, and not using QGIS to read arrow files is essential for
performance. Further performance was gained by not iterating over the
layer features and updating them, but by setting them directly based on
the fid (order). Note that the memory copy does not use the same fids as
our geopackage layers, and you can't control them. The last factor of
performance was disabling the automatic undo functionality (or rather
creating 1 undo, instead of one for each feature) when updating the
layer.

---------

Co-authored-by: Martijn Visser <mgvisser@gmail.com>