Linear allocation objective (#2174)
Fixes https://github.com/Deltares/Ribasim/issues/1494.

The objective function I have implemented now is of the form

$$
\sum_{i \in I} U_i + L_i,
$$

as described in the issue. I have reformulated the associated
constraints as

$$
d_i U_i \ge F_i - fd_i, \quad d_i L_i \ge d_i f- F_i
$$

to avoid division problems when $d_i = 0$. The 'target fraction' $f$ is
given by the minimum of 1 and the total capacity divided by the total
demand.

As mentioned by @jarsarasty, it could be that we have to add another
optimization in goal programming fashion to allocate the capacity that
is left after this first optimization. That is because the a priori
computed fraction $f$ as described above is only ideal if there are no
constraints between sources and demands.

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: Martijn Visser <mgvisser@gmail.com>
Co-authored-by: Maarten Pronk <git@evetion.nl>
Co-authored-by: visr <4471859+visr@users.noreply.github.com>
Co-authored-by: GitHub <noreply@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>