Reduce allocations in residual function (#2234)
Continuing from #2224:

```
18.086399 seconds (4.01 M allocations: 3.255 GiB, 2.81% gc time)
17.879659 seconds (2.95 M allocations: 2.949 GiB, 1.69% gc time)
```

Adding the type assert on the function does the trick. I also go rid of
the `isdae` check because it shouldn't be needed, and I noticed
`get_dae_uprev` leads to runtime dispatch since it is using
`isdefined(integrator, ...)`

There are still more optimizations to be made to backtracking, but good
to bag this one already.
I also noticed many extra allocations if `concentration = true`, we
should also look at that.