Reduce residual calls (#2242)
A follow-up to https://github.com/Deltares/Ribasim/pull/2224. I managed
to reduce the number of residual calls in the 'managing' of the
backtracking.

I cannot find any type instabilities in the relaxation code with Cthulhu
anymore. The profile with `hws_2025_4_0` now looks like this:


![image](https://github.com/user-attachments/assets/22c89dbc-2315-4121-a0f1-39da2b2b9a94)

What I also find interesting, is that time in the backtracking algorithm
is completely dominated by the first call to the residual, which I think
tells us that most of the time the residual reduction criterion is
immediately satisfied. Maybe we can come up with a cheap criterion in
the monitoring to skip the backtracking all together.

ping @visr @Huite