Refactor parameter initialization (#2251)
Fixes https://github.com/Deltares/Ribasim/issues/2133.

This refactor makes parameter initialization more explicit and I think
therefore easier to understand. It is also more type stable, because the
type of each parameter is obtained from the fields in the
`AbstractParameterNode` objects.

It is possible that this refactor changes untested behavior. One place
where I noticed that the order of parsing matters is with
`DiscreteControl`: the control state simulation starts out with depends
on it. I had to change the initial control state to the last control
state parsed to make all tests pass (although the failing test was a
test model where the controlled flow depends on itself which I think is
rare in practice).