Decrease relative tolerance over time. (#2277)
As an alternative to #2238. Could fix/workaround #2265 and #1897.

After some discussion with @SouthEndMusic, we believe that the water
balance error is purely a floating point error. Besides, it's always
extremely small, smaller than our tolerances, so it seems a red herring.

So that leaves the thing we keep twiddling with, the tolerances and our
cumulative flows, which are of course problematic, in the simple sense
that a flow of 1 by a thousand times decreases the effective relative
tolerance by 3. Instead of making the flows non (or at least less)
cumulative, or keep decreasing our default tolerances, this PR decreases
the relative tolerance over time, trying to compensate for the
accumulation. ~Ideally we change the calculation of the residual, but
since there's no easy way to [override
it](https://github.com/SciML/DiffEqBase.jl/blob/9c9733789a2c58562cf13340acf89d95d3cd1916/src/calculate_residuals.jl#L9-L11),
we change the tolerance.~

This should be tested with some of the problematic models described by
@visr, but first signs look good. An example, the hws model always
solved faster nearer to the end of the simulation, which can be
explained by the ever more relaxed tolerances. With this change the
computation time stays stable. The HWS 2025_4_0 model changes relative
tolerance three times:

```julia
[ Info: Relative tolerance changed at t = 259200.0 to 1.0e-8
[ Info: Relative tolerance changed at t = 2.16e6 to 1.0e-9
[ Info: Relative tolerance changed at t = 1.49472e7 to 1.0e-10
```

With these changes, I've lowered the default tolerances to 1e-5 without
problems.

The relative tolerances are lowered via a function callback at
exponentially increasing intervals. The relation is log10, but I found
that it's better to call it more often to keep the allowed error more
smooth:

<img width="738" alt="Screenshot 2025-05-21 at 09 37 14"
src="https://github.com/user-attachments/assets/ef0e9565-5410-47d3-b2c1-e601a6db378a"
/>

The formula for the new tolerance can be expressed as:

```math
\text{reltol'} = \max\left(10^{\log_{10}(\text{reltol}) - \text{diff\_norm}}, 10^{-14}\right)
```

With:

- $\text{reltol}$ is the configured relative tolerance.
- $`\text{diff\_norm} = \max\left(0,
\log_{10}\left(\frac{\text{cum\_magnitude}}{\text{avg\_magnitude}}\right)\right)`$
- $`\text{cum\_magnitude}`$ is the cumulative magnitude
($\text{internalnorm}$) of the state.
- $`\text{avg\_magnitude} = \max\left(\text{internalnorm}(10^4,
t),\frac{\text{cum\_magnitude}}{t}\right)`$ where $t$ is the current
time.

This formula ensures that the new tolerance decreases based on the
difference between the cumulative magnitude and the average magnitude
(once the cumulative magnitude exceeds $10^{4}$), with a lower bound of
$10^{-14}$ to maintain numerical precision.