Use tasks for s3 download and upload (#2301)
So we can use them as depends-on tasks in other workflows.

To be able to run the integration and regression tests locally with
these tasks:

```
pixi run test-ribasim-regression
pixi run model-integration-test
```

Developers now need to go to https://s3-console.deltares.nl/ to set up
personal access keys.
Make a copy of `.env.default` called `.env` and fill in the access key
and secret key there. That will allow us to download from MinIO the
reference data we need.

---------

Co-authored-by: Maarten Pronk <git@evetion.nl>