Allow Plotly 6 (#2303)
Locally this works to fix #2033 but we'll need to double check the docs
after merging this.

The errors fixed in 6.1 look very much like what we experienced:
https://github.com/plotly/plotly.py/blob/main/CHANGELOG.md#610---2025-05-15

I ran `pixi update plotly`. But if it causes to many conflicts we can
also just merge the unpinning in the TOML and let our monthly update
update it.

EDIT: let's just remove the pin and trust this fixes the issue.