QGIS reload Ribasim model (#2307)
This PR adds a Reload Ribasim model as right-click option in the layers
menu.

The QGIS plugin (and the dataset widget specifically) currently only
tracks the last opened model, which makes it hard to have a reload
button
work for other models. I often work with multiple models at the same
time, and I want to be able to reload any of them quickly.

This adds the full path as a custom property to all Ribasim made layers,
which is how we can detect a Ribasim group.
When clicked, we save the group position in the layer list, remove the
group, and then open the model again in the same position.
Any changes made to the group will thus be lost.

I've also added the full path as the group name, so that it is easier to
find the model in the list of layers.

<img width="578" alt="Screenshot 2025-05-29 at 09 05 10"
src="https://github.com/user-attachments/assets/5ea29d81-b7ba-49b1-8ed1-b51e096d3ed1"
/>