[TC] Try to use cache with --check-bounds=yes. (#2332)
Only some builds use the cache on TC, so this removes spaces from the
name (not allowed as indicated by trying to edit the cache feature in
the GUI). It also seems to recompile a lot, let's add --check-bounds=yes
to the instantiate (used in cache creation) to try to prevent it.