Convert docs from ipynb to qmd (#2376)
Needed for #2078, since that doesn't work on notebooks directly. Taking
this as an intermediate step to avoid getting to many conflicts.

Even if we don't make the switch, I think it is better to have all of
our docs in the same format. QMD support seems a bit more first class
and transparent when working in a repository.

I ran this locally and verified that the converted pages still look the
same.