Generate a Ribasim-python SBOM (#2342)
Similar to the julia SBOM generation, I've generated an SBOM via
sbom4python. The only thing I've done is pass on my module `-m ribasim`,
and it's still a bit magic to me which packages are then picked up. I
will have to learn more from the documentation, but at least there is an
SBOM with information.

---------

Co-authored-by: deltamarnix <deltamarnix@users.noreply.github.com>
Co-authored-by: Martijn Visser <mgvisser@gmail.com>