Explicit imports, complete the root project (#2392)
This fixed some issues highlighted by
https://github.com/JuliaTesting/ExplicitImports.jl:
- remove unused imports
- switch to explicit imports; don't rely on e.g. `using Tables`

Also so far we've made sure that all core and core test dependencies are
directly available in the root environment as well, for convenience.
This brings that back up to date.

Lastly, this switches TestItemRunner back to the main branch since
Maarten's PR is merged there.