Add convergence and dt to output. (#2405)
This should help with debugging, as we now only get the convergence
(`flow_error`) values on a crash/interrupt/end. Now we can see it over
time. This includes the `dt` in the solver stats. All values are
instantaneous for the saveat timestep.

It already gives interesting patterns for the `basic` model, such as a
peak later in time. There are similar peaks for the ManningResistance
after a month without a clear explanation.

<img width="833" alt="Screenshot 2025-06-26 at 09 09 36"
src="https://github.com/user-attachments/assets/ddb03cb7-8c84-4b65-b0e4-041da329812c"
/>

---------

Co-authored-by: Martijn Visser <mgvisser@gmail.com>