Improve Pixi dev doc (#2409)
Addresses points from #2408.

This also runs `initialize-julia` as part of `pixi run install`. We
don't run this task on CI (that is `install-ci`), and only advertise it
in the dev docs.

It will take quite a bit longer due to precompiling the Julia root env,
but since it is the last task you can interrupt it if needed.