Report cause of infeasibility (#2398)
Fixes #2328 

The cause of the infeasibility is determined by solving the same
optimization problem, with relaxed constraints.
The violation of the constraints is minimized, and so we find the least
violating constraint to make the problem feasible.

Next we check which variables contribute in this constraint, and report
all other constraints that also contain these variables. This way we get
a set of possible conflicting constraints, and log the value of each
variable.

We only relax the volume- and flow conservation constraints. This makes
sure that we immediately know at which node/edge the constraints are
violated.

---------

Co-authored-by: Bart de Koning <74617371+SouthEndMusic@users.noreply.github.com>
Co-authored-by: Jesús A. Rodríguez <jesusandres.rodriguez@deltares.nl>
Co-authored-by: Martijn Visser <mgvisser@gmail.com>