Use averaged normalized convergence error (#2448)
Instead of instantaneous at saveat, and a random atmp state at
interrupt. Solves #2414.

```
^C┌ Warning: Simulation crashed or interrupted at 2017-01-05T15:06:46.422.
└ @ Ribasim /Users/evetion/code/Ribasim/core/src/main.jl:42
┌ Warning: Convergence bottlenecks in descending order of severity:
│   ManningResistance #6409 = "8.03%"
│   ManningResistance #6452 = "4.99%"
│   ManningResistance #6708 = "1.81%"
│   ManningResistance #6542 = "0.99%"
│   ManningResistance #6443 = "0.95%"
```

This is the averaged normalized error. At 100% it would be the node that
had the maximum error (normalized to 1) at the end of each calculation
timestep. When the error is random, each node would have the same
convergence percentage of 1/length(u).

---

I've investigated overriding `internalnorm`, so we can catch `atmp`
inside of the nonlinearsolves, but it turned out to be too hacky,
probably requiring `stacktrace()` just to check whether it's been called
from a solver, instead of some random other pieces of machinery.