Use pandera v0.25, and revert using pyarrow dtypes (#2450)
This fixes #2120, we can update to the latest Pydantic release (0.25)
again. There is still one open bug
https://github.com/unionai-oss/pandera/issues/2016. That triggered me to
essentially revert #1781 and switch back to the default pandas dtypes,
because with them we are less likely to hit various compatibility issues
in the ecosystem.

Because we like nullable strings, integers and Booleans, we do use these
now:
- Nullable integer data type:
https://pandas.pydata.org/docs/user_guide/integer_na.html
- Nullable Boolean data type:
https://pandas.pydata.org/docs/user_guide/boolean.html
- StringDtype: https://pandas.pydata.org/docs/user_guide/text.html

Furthermore this gets rid of Series in the type annotations using
https://pandera.readthedocs.io/en/stable/dataframe_models.html#using-data-types-directly-for-column-type-annotations.
This was needed because otherwise MyPy would complain about not allowing
certain dtypes, but it is cleaner anyway.

This also switches back to the default Timestamp with ns precision for
the same reason. Possibly we need to support other units in the future
if we do very long simulations.

---------

Co-authored-by: Maarten Pronk <git@evetion.nl>