Improve continuous control performance (#2446)
Fixes https://github.com/Deltares/Ribasim/issues/2445.

The model `repro-models/HollandseDelta_parameterized` makes heavy use of
ContinuousControl. I had the hypothesis that the fact that the functions
that define the continuous relations for ContinuousControl are piecewise
linear is hard on the solver due to the lack of smoothness. To test this
I changed the interpolation type to `PCHIPInterpolation` (see
[here](https://docs.sciml.ai/DataInterpolations/stable/methods/#PCHIP-Interpolation)),
not as a change to merge, but just to see what a smoother interpolation
type does. And indeed, with this smoother interpolation I can locally
run the first 4 months of the simulation in 9 min.

There are several things we can do to improve the ContinuousControl
performance structurally:
- Indeed use a smoother interpolation type than LinearInterpolation. We
of course have to use an interpolation type that can be understood by
modelers, so this could be a good use case for [smoothed linear
interpolation](https://southendmusic.github.io/CachedInterpolations.jl/stable/examples/#Smoothed-linear-interpolation)
- Treat ContinuousControl more similar to DiscreteControl, in the sense
that the continuous control logic is not part of the ODE system, but the
control logic is computed at the start of the timestep and then kept
fixed over the timestep. This approach probably has the best performance
gain, but also results in a loss of accuracy. Alternatively we could
make the continuous-but-not-part-of-the-ODE-system-control a new node
type (or add a switch in the TOML between callback computed or ODE
computed), but then we need to make the subtle difference very clear in
the docs.

@visr @gijsber @Huite let me know whether you think this warrants a
design session. Maybe in the particular case of the `HollandseDelta`
model the schematization can also be made a bit more solver friendly

As a side node, I think we can expect similar issues with models that
have a lot of TabulatedRatingCurve nodes with very sharp transitions in
the q(h) relations