Support giving `model_dir` to `run_delwaq`
Right now `run_delwaq` doesn't take arguments and assumes that the model is in a model dir next to the ribasim.delwaq folder. To make it possible to run models in other places this adds a kwarg `model_dir`, with the same default.

Perhaps we should make the kwarg mandatory, but I don't want to break it right now.