QGIS: start removing editing features
This removes creating new models, creating link geometries automatically, suppressing attribute forms, and the "Add to QGIS" and "Remove from Dataset" buttons.

The Ribasim widget could probably be entirely removed and just replaced with an open model dialog, but that is a bit more involved since we hang things on the widgets that we need for e.g. results.