Remove Legolas, take 2
This addresses the parts of #24832 that I wasn't happy with. Now schema.jl contains the Schema module, which has a module for each node type. Each node type module then has a struct for each table schema. So `Schema.Basin.State` is now a struct. We use the type of this thing to load the right table, and the resulting table will be a `StructVector{Schema.Basin.State}`, meaning when you iterate over the rows you get these structs filled with data.

Fixes #1127.