LevelDemand in terms of absolute error (#2515)
Fixes https://github.com/Deltares/Ribasim/issues/2505

With this PR the LevelDemand is defined in a very straightforward way.
Given the target levels $h_\min^p$ and $h_\max^p$ for priority $p$, the
storage at the end of the allocation timestep $S_\text{end}$ the profile
$s(h)$ and the absolute error $E$, we have:

$$
   E \ge s(h_\text{min}^p) - S_\text{end}
$$

$$
   E \ge S_\text{end} - s(h_\text{max}^p)
$$

which is well defined independent of the state of the model. This does
suffer from uniqueness problems, because the error reduces the same for
very Basin when some storage is allocated to it. As far as I can see,
the cleanest way to obtain an error formulation that yields uniqueness
of the optimal solution is to use quadratic errors.

I did another simplification in this PR: For all demand nodes there is
now a single error term, instead of separate upper and lower error
terms.

@jarsarasty