Avoid `Tables.columntable` on `SQLite.Query` (#2533)
This rolls our own implementation, which is much faster. Here is the
before and after of initializing a model with a large database:

```
# 296.342815 seconds (331.38 M allocations: 27.221 GiB, 24.09% gc time, 0.03% compilation time)
# 93.071874 seconds (385.77 M allocations: 29.021 GiB, 16.68% gc time, 0.07% compilation time)
```

Especially Basin / time was large. I will move that table out to an
Arrow table for all LHM models to speed up initialization, but this
still seems worth it. It also avoids loading extra meta columns in
memory.