Set juliacall env vars for quarto-preview as well (#2543)
Doing this in a fresh clone caused the root Project.toml to be
overwritten; see also issue
https://github.com/JuliaPy/pyjuliapkg/issues/62.

That lead to errors like:

```
JuliaError: ArgumentError: Package Ribasim not found in current path.
```

Should fix the issue that @simulutions and @Fati-Mon ran into. I tested
this in a fresh clone and there it was now fixed.