Add `pixi run filter-model` (#2551)
This utility can be run to avoid some of the pain from
https://github.com/Deltares/Ribasim/issues/2550.

It removes all data outside the simulation period. Note that if there is
no timestamp on the starttime or endtime this can affect results or
break the model. It is just meant as a convenience function for
developers, nothing public.

Output looks like this:
```
Filtering time tables to period: 2023-01-20 00:00:00 to 2024-04-27 00:00:00
Basin / time                        removed 6774500 / 7760500 rows (87.3%)
FlowBoundary / time                 removed 45 / 973 rows (4.6%)
LevelBoundary / time                removed 432 / 1360 rows (31.8%)
Total rows removed across all time tables: 6774977
```

I also modified both `filter-model` and `migrate-model` to accept one
argument, in which case it will modify the model in place, use with
caution.