Test source priorities (#2553)
Fixes https://github.com/Deltares/Ribasim/issues/2051.

The test model looks like this:

<img width="629" height="471" alt="image"
src="https://github.com/user-attachments/assets/79721a86-1f40-4548-bc7d-e65e041837de"
/>

The idea is that the UserDemand gradually increases its demand over a
single priority. The level boundaries go from more desirable to less
desirable to the right. We see that this is correctly reflected in the
flows over the pumps:

<img width="588" height="394" alt="image"
src="https://github.com/user-attachments/assets/4372e356-2615-4845-ac8e-df69057842c2"
/>

I did have to change the objective for the source priorities because it
did exactly the opposite of what we want. The objective function to
minimize is now

$$
\sum w F
$$

where $F$ is the flow out of a source and $w$ is its source priority.
This means that the higher the source priority value (which means the
lower the desirability of the source) the more the solver is
incentivized to lower $F$.