Fix CLI panic for threads arg (#2567)
I should've triggered the CLI tests manually on #2562.

It failed with
```
thread 'main' panicked at src/main.rs:50:55:
called `Option::unwrap()` on a `None` value
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
```

The problem was that when neither `--threads` nor `JULIA_NUM_THREADS`
was set, it tried to unwrap `cli.threads` which was none.

This avoids that case and adds some tests.