Implement hysteresis on DiscreteControl. (#2564)
Possibly fixes #2444, as requested by @ngoorden and @DanielTollenaar.

This introduces a secondary threshold to the condition table of
DiscreteControl, enabling the definition a bandwith (deadband), which is
used for hysteresis. `greater_than` is renamed to `threshold_high`, and
the secondary threshold is named `threshold_low`. The high threshold is
used when the condition was false in a previous timestep, while the
`threshold_low` is used when the condition was true in the previous
timestep. You can thus get the following behaviour (in Dutch, credit
@ngoorden from their powerpoint), with the `threshold_high` set to 0.95
m, and `threshold_low` set to 0.9 m. Within the deadband, the control
state retains the previous control state.

<img width="622" height="282" alt="image"
src="https://github.com/user-attachments/assets/9ed47885-c75c-4354-8f22-89921a09b28a"
/>

The `threshold_low` column is optional, and by default set to the value
of `threshold_high`, thus disabling hysteresis.

PS. We could also call the thresholds upper and lower?

---------

Co-authored-by: lexverheem <lex.verheem@deltares.nl>
Co-authored-by: Martijn Visser <mgvisser@gmail.com>