Support NetCDF input tables as Arrow alternative (#2542)
Fixes most of #2434, not all because only a subset of the tables are
supported.

This essentially gives the same treatment we currently have for Arrow
input files, but for NetCDF files. We choose between the file format
based on whether the path has a `.arrow` or `.nc` extension.

So in Python if you do:
```py
model.basin.state.set_filepath(Path("basin-state.nc"))
```
(Note you need
[this](https://github.com/Deltares/Ribasim/issues/2039#issuecomment-3204851989)
if you read the model from disk.)

Ribasim Python will convert the `df.to_xarray().to_netcdf()`, and if you
read it back in with Python `ds.to_dataframe()`.
In the TOML you get:
```toml
[basin]
state = "basin-state.nc"
```

The core then also reads based on the file extension. To keep things
simple, when reading data into the core we convert the N dimensional
arrays in the NetCDF to tables, so the rest of the initialization
doesn't need to handle both data structures.

We should also think about what to do for irregular tables that don't
fit in a structured array. Perhaps just throw an error?