Reset `ribasim_version` in `model_post_init` (#2578)
I noticed that when reading in old models with Ribasim Python, with an
old `ribasim_version` in the TOML, this version was preserved by Ribasim
Python. This isn't correct, since when we read in the model, we run the
migration code automatically, so we should reset this key.

This does that, and tests that it is set in memory and written to the
TOML.