Fix empty Arrow results (#2585)
Fixes #2580.

The QGIS Windows standalone installer of our minimum supported QGIS 3.34
does come with pyarrow installed. I tested with 3.14.15, and the
reported issues were all with older patch releases. And I see in issues
that some older patch releases did have pyarrow related issues.

But as mentioned in
https://github.com/Deltares/Ribasim/issues/2580#issuecomment-3302910110,
we have a fallback for missing pyarrow, that we don't test. It was
broken for empty Arrow tables, since `data` was None in this case,
giving us a DataFrame with not only 0 rows (fine) but also 0 columns
(bad), which triggered the stacktrace in the issue. This case is now
explicitly handled.

After fixing that I ran into another issue, that columns were missing
from the allocation result files when allocation was not used. It turned
out that recently switching to `getfield.(v, :node_id)` will give us
`Any[]`. This is fixed by using StructArrays to handle these Array Of
Structs, so we can just use `v.node_id` and get `Int32[]`. To make sure
this doesn't regress we now also test the schemas of the empty tables.