Update typing in Ribasim Python (#2581)
I asked Claude to go over the type annotations, and also update to the
syntax supported by our minimum Python version 3.11.

It suggested in several places to use `object` instead of `Any`, the
latter bails out of typing. Also it could refine the types of our
context variables and generator.

---------

Co-authored-by: Maarten Pronk <git@evetion.nl>