QGIS: Always use GDAL instead of pyarrow (#2591)
We missed https://github.com/Deltares/Ribasim/issues/2580 before the
release because all developers have pyarrow in their QGIS, but many/most
users do not. So the fallback path was not well tested.

This promotes the old fallback to the only way to load data. I was
worried that it would be slower, but I timed both paths to read
`lhm_coupled_2025_9_0`:

> PyArrow feather read took 0.4215 seconds
> PyArrow feather read took 0.4256 seconds

> OGR arrow stream read took 0.4783 seconds
> OGR arrow stream read took 0.4756 seconds

Close enough.

The behavior of the GDAL code is slightly different. Instead of loading
empty tables, this doesn't load the layers. I think that is better. So
e.g. for this model without allocation the allocation results don't show
up in the layers.