Use quarto-publish task (#2594)
The action in https://github.com/Deltares/Ribasim/pull/2593 cannot find
quarto, and this does not seem configurable. So it is failing on main.

Since we run render the step before, I call `--no-render`.
This uses git worktrees and when trying out this task locally pre-commit
failed on gh-pages which doesn't have a pre commit yaml, hence
PRE_COMMIT_ALLOW_NO_CONFIG.

The `_publish.yml` is created by quarto, and the docs seem to indicate
this needs an ID, but I haven't been able to generate that yet.