Support Scientific Python SPEC 0 (#2630)
Notably this moves our default Python to 3.13, and drops 3.11 support.

We can bump most deps.
Xarray was recently added as a dep and I set compat to a recent version,
but we should support until 2023.11 at the moment. We don't test it, but
AFAIK it should work.

https://scientific-python.org/specs/spec-0000/

Drop QGIS dep on macOS: We haven't been able to move to Python 3.13 for
the longest time, and there is still no macOS build. I propose we drop
it and move on. @evetion does that cause issues for you?